requirements.txt
absl-py==2.1.0
aiosignal==1.3.1
attrs==22.2.0
backcall==0.2.0
brotlipy==0.7.0
click==7.1.2
cloudpickle==1.2.2
cycler==0.11.0
Cython==0.29.28
decorator==5.1.1
deepdiff==6.7.1
dm-env==1.6
dm-env-rpc==1.1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.1
enum34==1.1.10
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.12.2
frozenlist==1.3.3
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.5.1
googleapis-common-protos==1.69.2
graphviz==0.20.1
grpcio==1.62.3
gym==0.14.0
gymnasium==0.28.1
imageio==2.26.1
immutabledict==2.2.5
importlib-metadata==6.7.0
importlib-resources==1.5.0
ipdb==0.13.13
ipython==7.34.0
jax[cpu]==0.4.31
jaxlib
jaxmarl
jax-jumpy==1.0.0
jedi==0.19.2
Jinja2==3.0.3
jsonschema==4.17.3
kiwisolver==1.4.5
lxml==5.3.2
MarkupSafe==2.1.5
matplotlib==3.0.3
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-service==2.4.0
mock==5.2.0
mpyq==0.2.5
msgpack==1.0.5
msgpack-numpy==0.4.4.3
mujoco==2.3.6
mujoco-py==2.1.2.14
networkx==2.6.3
ordered-set==4.1.0
packaging==24.0
pandas==0.25.3
parso==0.8.4
patchelf==0.17.2.2
PettingZoo==1.22.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
portpicker==1.6.0
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pygame==2.1.2
pyglet==1.3.2
Pygments==2.17.2
pymunk==6.2.1
PyOpenGL==3.1.9
pyparsing==3.1.4
pyrsistent==0.19.3
PySC2==4.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
ray==2.7.2
recordtype==1.4
s2clientprotocol==5.0.14.93333.0
s2protocol==5.0.14.93333.0
scipy==1.7.3
sentry-sdk==2.25.1
setproctitle==1.3.3
six==1.17.0
sk-video==1.1.10
SMAC @ git+https://github.com/oxwhirl/smac.git@d6aab33f76abc3849c50463a8592a84f59a5ef84
smmap==5.0.2
SuperSuit==3.7.0
svgutils==0.3.4
timeout-decorator==0.4.1
tinyscaler==1.2.7
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.67.1
traitlets==5.9.0
typing_extensions==4.7.1
wandb==0.18.7
wcwidth==0.2.13
websocket-client==1.6.1
yarg==0.1.10
zipp==3.15.0
